HvD: Parallelize the evaluation of the functional over multiple GPU cores.
Note that the code uses an interesting trick to check which derivatives need
to be computed. They use the order of derivative of the density, and
if a derivative is not to be computed the pointer to the corresponding
memory should be NULL. However, we need to invoke every functional
evaluation with the memory associated with the corresponding grid point.
Hence we need a second set of "helper" variables to program around the
potential NULL pointers.