HvD: Addressing multi-term functionals.
Not completely working yet (synchronization?). For multi-term functionals
twice as much memory is allocated for the results. The second half of
the memory is used internally to buffer the results from the individual
terms. The BLAS routine cublasDaxpy is used to add the results to the
functional. At the moment some results are correct and some aren't.
Nevertheless, I need to commit as my account on Ascent is going to
disappear later today. At least the changes don't break anything that
worked before.