HvD: Finishing addition of streams.
Using variadic functions in C failed because there is no good way of
telling the function being called of how many optional arguments are
provided in the call. In addition there is also no way of telling
the function what the data types are. Variadic functions really only
work for cases that are very much to printf, where there is some
argument, like the format string, that can be used to communicate the
number and types of the optional arguments. For cases like LibXC there
is no good way to use this facility.

Hence the solution is to modify the API so you always have to pass a
stream argument. If you don't want to use streams then you have to
pass the zero stream. The downside is that the API for non-CUDA
platforms would get polluted with CUDA specific arguments. This
problem is avoided by using ifdef-s to strip these arguments out.
The fact that the API changes when using a non-CUDA compiler is not a
problem because the corresponding implementation is also stripped
out and so those parts cannot be used in any case.