HvD: Replace cudaMallocManaged with regular host and device memory and explicit data copies
The cudaMallocManaged memory in the unified address space only works on the
very latest bleeding edge hardware and software. Using this in a regression
test of an open source user library is not a good idea as it breaks on most
platforms. Hence, I am replacing the usage of this capability with the good
old fashioned way of doing things. Once this new capability is supported
widely we can always switch back to it.